Fix failing unit tests, update version number
When building as a shared library, one of the unit tests was failing.
Specifically, it was writing out the nodes in a different order than the
test expected. Since the order isn't defined, the test was actually
fine. So we check if the node is written out in one of two possible
orders.

Also, gmock and gtest were linked the to unit tests. Running the unit
tests when yaml-cpp was built as a shared library caused a free of
already deallocated memory.  This was fixed by only linking against
gmock.